+++ R/check.R
@@ -13,6 +13,7 @@ check = function(server = NULL, token = NULL,
                  dir = ".", file = "config-uat.yml") {
   summarise_setup(server, token)
   user_details = summarise_user(get_server(), get_token())
+  check_server_version(get_server(), get_token())
   summarise_versions(get_server(), get_token())
 
   register_uat_user(get_server(), get_token(), account = user_details$username)