+++ R/check_server.R
@@ -1,0 +1,34 @@
+check_server_version = function(server, token) {
+  cli::cli_h2("Checking Server Version")
+  con = suppressMessages(connectapi::connect(server = server, api_key = token))
+
+  server_version = con$server_settings()$version
+  connect_versions = get_connect_versions()
+  row_number = which.max(connect_versions$version == server_version)
+
+  # Version not in the database!
+  if (length(row_number) == 0L) row_number = nrow(connect_versions) + 1
+
+  # Up to date - edge case :)
+  if (row_number == 1L) {
+    cli::cli_alert_info("Your server is up to date")
+    return(NULL)
+  }
+
+  connect_versions = connect_versions[seq_len(row_number - 1), ]
+  no_of_versions = length(unique(connect_versions$version)) #nolint
+  cli::cli_alert_info("Your server is {cli::col_red('out of date')}")
+  cli::cli_alert_info("There {cli::col_red(no_of_versions)} newer versions that fix \\
+                      {cli::col_red(nrow(connect_versions))} CVEs")
+}
+
+get_connect_versions = function() {
+  versions = system.file("extdata", "connect_versions", "connect_versions.csv",
+                         mustWork = TRUE, package = "jrHealthCheckConnect")
+  versions = utils::read.csv(versions)
+  versions$date = as.Date(versions$date)
+  versions = dplyr::tibble(versions)
+
+  versions = dplyr::arrange(versions, dplyr::desc(date))
+  return(versions)
+}